requirements.txt
flask==2.0.2
pillow==8.2.0
numpy==1.21.0
tensorflow
